Break the cycle between SMTLowerer and SMTLoweringRewritePattern
