+++ xdsl_smt/passes/lower_to_smt/func_to_smt.py
@@ -7,8 +7,8 @@ from xdsl.dialects.func import FuncOp, Return
 from xdsl_smt.dialects.effects.effect import StateType
 from xdsl_smt.passes.lower_to_smt.smt_lowerer import (
     SMTLowerer,
-    SMTLoweringRewritePattern,
 )
+from xdsl_smt.passes.lower_to_smt.smt_rewrite_patterns import SMTLoweringRewritePattern
 from xdsl_smt.dialects.smt_dialect import DefineFunOp, ReturnOp
